refactor(models): migrate Fibex topology + Fibex4Ethernet family per spec packages (Rule 0007 A)
- Move SocketConnection/SoAdRoutingGroup to ObsoleteModel (R4.3.1 package, obsolete in R23-11)
- Extract TcpOptionFilterSet and IPv6HeaderFilterList into standalone modules
- Move RequestResponseDelay EthernetTopology -> ServiceInstances (R23-11 Table 6.171)
- Move TransportProtocolConfiguration family ServiceInstances -> EthernetTopology (R23-11 Tables 6.125-6.133)
- Relocate physical channels out of CoreTopology (Ethernet/Flexray/Lin/Can topology modules)
- Fix circular imports via TYPE_CHECKING guards and deferred annotation evaluation
- Annotate parity checklists with verified spec release (R4.3.1 vs R23-11) and table numbers
