+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/FibexCore/CoreTopology.py
@@ -4,15 +4,29 @@ from abc import ABC
 from typing import TYPE_CHECKING, List, Optional
 
 if TYPE_CHECKING:
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanCommunication import CanFrameTriggering
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanTopology import CanPhysicalChannel
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import NetworkEndpoint
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanTopology import (
+        CanClusterBusOffRecovery,
+        CanCommunicationConnector,
+        CanCommunicationController,
+        CanPhysicalChannel,
+    )
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import (
+        EthernetCommunicationConnector,
+        EthernetCommunicationController,
+        EthernetPhysicalChannel,
+    )
     from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Flexray.FlexrayCommunication import FlexrayFrameTriggering
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinCommunication import LinFrameTriggering, LinScheduleTable
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanTopology import CanCommunicationConnector, CanCommunicationController
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import EthernetCommunicationConnector, EthernetCommunicationController
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Flexray.FlexrayTopology import FlexrayCommunicationConnector, FlexrayCommunicationController
-    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinTopology import LinCommunicationConnector, LinMaster
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Flexray.FlexrayTopology import (
+        FlexrayCommunicationConnector,
+        FlexrayCommunicationController,
+    )
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinCommunication import LinFrameTriggering
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinTopology import (
+        LinCommunicationConnector,
+        LinMaster,
+        LinPhysicalChannel,
+    )
+    from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanCommunication import CanFrameTriggering
     from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreCommunication import (
         CommunicationDirectionType,
         FramePort,
@@ -275,126 +289,6 @@ class PhysicalChannel(Identifiable, ABC):
         return self.getElement(short_name, PduTriggering)
 
 
-class LinPhysicalChannel(PhysicalChannel):
-    """
-    Represents a LIN physical channel in the communication system,
-    defining LIN-specific properties including bus idle timeout
-    and schedule tables for LIN network communication.
-    """
-
-    # LinPhysicalChannel method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getBusIdleTimeoutPeriod      [x] impl  [ ] docstring  [ ] test
-    # [ ] setBusIdleTimeoutPeriod      [x] impl  [ ] docstring  [ ] test
-    # [ ] getScheduleTables            [x] impl  [ ] docstring  [ ] test
-    # [ ] createLinScheduleTable       [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.busIdleTimeoutPeriod: TimeValue = None
-        self.scheduleTables: List[LinScheduleTable] = []
-
-    def getBusIdleTimeoutPeriod(self):
-        return self.busIdleTimeoutPeriod
-
-    def setBusIdleTimeoutPeriod(self, value):
-        if value is not None:
-            self.busIdleTimeoutPeriod = value
-        return self
-
-    def getScheduleTables(self):
-        return self.scheduleTables
-
-    def createLinScheduleTable(self, short_name: str) -> LinScheduleTable:
-        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinCommunication import LinScheduleTable
-
-        if not self.IsElementExists(short_name, LinScheduleTable):
-            end_point = LinScheduleTable(self, short_name)
-            self.addElement(end_point)
-            self.scheduleTables.append(end_point)
-        return self.getElement(short_name, LinScheduleTable)
-
-
-class VlanConfig(Identifiable):
-    """
-    Defines Virtual LAN (VLAN) configuration properties,
-    specifying VLAN identifiers for network segmentation
-    and traffic management in Ethernet communication.
-    """
-
-    # VlanConfig method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getVlanIdentifier            [x] impl  [ ] docstring  [ ] test
-    # [ ] setVlanIdentifier            [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.vlanIdentifier: PositiveInteger = None
-
-    def getVlanIdentifier(self):
-        return self.vlanIdentifier
-
-    def setVlanIdentifier(self, value):
-        if value is not None:
-            self.vlanIdentifier = value
-        return self
-
-
-class EthernetPhysicalChannel(PhysicalChannel):
-    """
-    Represents an Ethernet physical channel in the communication system,
-    defining Ethernet-specific properties including network endpoints,
-    Socket Adaptor (SoAd) configuration, and VLAN settings.
-    """
-
-    # EthernetPhysicalChannel method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getNetworkEndpoints          [x] impl  [ ] docstring  [ ] test
-    # [ ] createNetworkEndPoint        [x] impl  [ ] docstring  [ ] test
-    # [ ] getSoAdConfig                [x] impl  [ ] docstring  [ ] test
-    # [ ] setSoAdConfig                [x] impl  [ ] docstring  [ ] test
-    # [ ] getVlan                      [x] impl  [ ] docstring  [ ] test
-    # [ ] createVlanConfig             [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.networkEndpoints: List[NetworkEndpoint] = []
-        self.soAdConfig = None
-        self.vlan: VlanConfig = None
-
-    def getNetworkEndpoints(self):
-        return self.networkEndpoints
-
-    def createNetworkEndPoint(self, short_name: str) -> "NetworkEndpoint":
-        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import NetworkEndpoint
-
-        if not self.IsElementExists(short_name, NetworkEndpoint):
-            end_point = NetworkEndpoint(self, short_name)
-            self.addElement(end_point)
-            self.networkEndpoints.append(end_point)
-        return self.getElement(short_name, NetworkEndpoint)
-
-    def getSoAdConfig(self):
-        return self.soAdConfig
-
-    def setSoAdConfig(self, value):
-        self.soAdConfig = value
-        return self
-
-    def getVlan(self):
-        return self.vlan
-
-    def createVlanConfig(self, short_name: str) -> VlanConfig:
-        if not self.IsElementExists(short_name, VlanConfig):
-            config = VlanConfig(self, short_name)
-            self.vlan = config
-            self.addElement(config)
-        return self.getElement(short_name, VlanConfig)
-
-
 class FlexrayChannelName(AREnum):
     """
     Enumeration defining names for FlexRay channels,
@@ -410,32 +304,6 @@ class FlexrayChannelName(AREnum):
 
     def __init__(self):
         super().__init__([FlexrayChannelName.CHANNEL_A, FlexrayChannelName.channel_B])
-
-
-class FlexrayPhysicalChannel(PhysicalChannel):
-    """
-    Represents a FlexRay physical channel in the communication system,
-    defining FlexRay-specific properties including channel name
-    designation for dual-channel FlexRay communication.
-    """
-
-    # FlexrayPhysicalChannel method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getChannelName               [x] impl  [ ] docstring  [ ] test
-    # [ ] setChannelName               [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        self.channelName = None  # type: FlexrayChannelName
-
-    def getChannelName(self):
-        return self.channelName
-
-    def setChannelName(self, value):
-        if value is not None:
-            self.channelName = value
-        return self
 
 
 class CommunicationCluster(FibexElement, ABC):
@@ -512,12 +380,16 @@ class CommunicationCluster(FibexElement, ABC):
         """
         This relationship defines which channel element belongs to which cluster. A channel shall be assigned to exactly one cluster, whereas a cluster may have one or more channels. Note: This atpSplitable property has no atp.Splitkey due to atpVariation (PropertySetPattern). Stereotypes: atpSplitable; atpVariation Tags: vh.latestBindingTime=systemDesignTime
         """
+        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinTopology import LinPhysicalChannel
+
         return list(sorted(filter(lambda a: isinstance(a, LinPhysicalChannel), self.physicalChannel), key=lambda o: o.getShortName()))
 
     def getEthernetPhysicalChannels(self) -> List[EthernetPhysicalChannel]:
         """
         This relationship defines which channel element belongs to which cluster. A channel shall be assigned to exactly one cluster, whereas a cluster may have one or more channels. Note: This atpSplitable property has no atp.Splitkey due to atpVariation (PropertySetPattern). Stereotypes: atpSplitable; atpVariation Tags: vh.latestBindingTime=systemDesignTime
         """
+        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import EthernetPhysicalChannel
+
         return list(sorted(filter(lambda a: isinstance(a, EthernetPhysicalChannel), self.physicalChannel), key=lambda o: o.getShortName()))
 
     def createCanPhysicalChannel(self, short_name: str):
@@ -536,6 +408,8 @@ class CommunicationCluster(FibexElement, ABC):
         """
         This relationship defines which channel element belongs to which cluster. A channel shall be assigned to exactly one cluster, whereas a cluster may have one or more channels. Note: This atpSplitable property has no atp.Splitkey due to atpVariation (PropertySetPattern). Stereotypes: atpSplitable; atpVariation Tags: vh.latestBindingTime=systemDesignTime
         """
+        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Lin.LinTopology import LinPhysicalChannel
+
         if not self.IsElementExists(short_name, LinPhysicalChannel):
             channel = LinPhysicalChannel(self, short_name)
             self.addElement(channel)
@@ -546,6 +420,8 @@ class CommunicationCluster(FibexElement, ABC):
         """
         This relationship defines which channel element belongs to which cluster. A channel shall be assigned to exactly one cluster, whereas a cluster may have one or more channels. Note: This atpSplitable property has no atp.Splitkey due to atpVariation (PropertySetPattern). Stereotypes: atpSplitable; atpVariation Tags: vh.latestBindingTime=systemDesignTime
         """
+        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import EthernetPhysicalChannel
+
         if not self.IsElementExists(short_name, EthernetPhysicalChannel):
             channel = EthernetPhysicalChannel(self, short_name)
             self.addElement(channel)
@@ -556,6 +432,8 @@ class CommunicationCluster(FibexElement, ABC):
         """
         This relationship defines which channel element belongs to which cluster. A channel shall be assigned to exactly one cluster, whereas a cluster may have one or more channels. Note: This atpSplitable property has no atp.Splitkey due to atpVariation (PropertySetPattern). Stereotypes: atpSplitable; atpVariation Tags: vh.latestBindingTime=systemDesignTime
         """
+        from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Flexray.FlexrayTopology import FlexrayPhysicalChannel
+
         if not self.IsElementExists(short_name, FlexrayPhysicalChannel):
             channel = FlexrayPhysicalChannel(self, short_name)
             self.addElement(channel)
@@ -590,76 +468,6 @@ class CommunicationCluster(FibexElement, ABC):
         """
         if value is not None:
             self.protocolVersion = value
-        return self
-
-
-class CanClusterBusOffRecovery(ARObject):
-    """
-    Defines bus off recovery properties for CAN clusters,
-    specifying timing and counter configurations for
-    CAN controller recovery after bus off conditions.
-    """
-
-    # CanClusterBusOffRecovery method parity checklist:
-    # [ ] __init__                     [x] impl  [ ] docstring  [ ] test
-    # [ ] getBorCounterL1ToL2          [x] impl  [ ] docstring  [ ] test
-    # [ ] setBorCounterL1ToL2          [x] impl  [ ] docstring  [ ] test
-    # [ ] getBorTimeL1                 [x] impl  [ ] docstring  [ ] test
-    # [ ] setBorTimeL1                 [x] impl  [ ] docstring  [ ] test
-    # [ ] getBorTimeL2                 [x] impl  [ ] docstring  [ ] test
-    # [ ] setBorTimeL2                 [x] impl  [ ] docstring  [ ] test
-    # [ ] getBorTimeTxEnsured          [x] impl  [ ] docstring  [ ] test
-    # [ ] setBorTimeTxEnsured          [x] impl  [ ] docstring  [ ] test
-    # [ ] getMainFunctionPeriod        [x] impl  [ ] docstring  [ ] test
-    # [ ] setMainFunctionPeriod        [x] impl  [ ] docstring  [ ] test
-
-    def __init__(self):
-        super().__init__()
-
-        self.borCounterL1ToL2: PositiveInteger = None
-        self.borTimeL1: TimeValue = None
-        self.borTimeL2: TimeValue = None
-        self.borTimeTxEnsured: TimeValue = None
-        self.mainFunctionPeriod: TimeValue = None
-
-    def getBorCounterL1ToL2(self):
-        return self.borCounterL1ToL2
-
-    def setBorCounterL1ToL2(self, value):
-        if value is not None:
-            self.borCounterL1ToL2 = value
-        return self
-
-    def getBorTimeL1(self):
-        return self.borTimeL1
-
-    def setBorTimeL1(self, value):
-        if value is not None:
-            self.borTimeL1 = value
-        return self
-
-    def getBorTimeL2(self):
-        return self.borTimeL2
-
-    def setBorTimeL2(self, value):
-        if value is not None:
-            self.borTimeL2 = value
-        return self
-
-    def getBorTimeTxEnsured(self):
-        return self.borTimeTxEnsured
-
-    def setBorTimeTxEnsured(self, value):
-        if value is not None:
-            self.borTimeTxEnsured = value
-        return self
-
-    def getMainFunctionPeriod(self):
-        return self.mainFunctionPeriod
-
-    def setMainFunctionPeriod(self, value):
-        if value is not None:
-            self.mainFunctionPeriod = value
         return self
